chore(refactor): refactor current OBT winner into component
